# stocktool-kiosk/openapi-cloud-api.yaml
openapi: 3.0.3
info:
  title: StockTool Cloud API
  description: |
    Inventory, tools, projects, and barcode management, plus the
    installation/sync/config/update surface used by the StockTool Kiosk
    (v2) desktop client.

    ## Authentication
    Two separate credential types are used depending on the caller:

    - **User JWT** (`Authorization: Bearer <jwt>`) — obtained from
      `POST /api/auth/login` or `POST /api/auth/kiosk-login`. Used by the
      admin frontend and any endpoint acting on behalf of a specific
      person. Admin-only endpoints additionally require the JWT's user to
      have `role: admin`.
    - **Installation token** (`Authorization: Bearer <token>`) — obtained
      once from `POST /api/installations/register`. Used by the kiosk-v2
      desktop client for `/api/config`, `/api/updates/latest`, and
      `/api/sync/*`. Represents a trusted *device* relaying many local
      users' actions, not one person's session — a different credential
      type from the user JWT on purpose.
  version: "1.0.0"
servers:
  - url: https://api-stocktool.opslabsystems.cloud
    description: Production
...
security:
  - userJwt: []
paths:
...
  /api/tools/:
    get:
      tags: [Tools]
      summary: List tools
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [available, checked_out, broken, under_repair, stolen, lost] } }
        - { name: category, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: overdue, in: query, schema: { type: string, enum: ["true", "false"] } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Tool" } } } } }
    post:
      tags: [Tools]
      summary: Create a tool (admin only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                tool_number: { type: string }
                brand: { type: string }
                model: { type: string }
                serial_number: { type: string }
                description: { type: string }
                category: { type: string }
                location: { type: string }
                purchase_date: { type: string, format: date }
                purchase_price: { type: number }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Tool" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/tools/{tool_id}:
    get:
      tags: [Tools]
      summary: Get a tool
      parameters: [{ $ref: "#/components/parameters/ToolId" }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Tool" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Tools]
      summary: Update a tool (admin only)
      parameters: [{ $ref: "#/components/parameters/ToolId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                tool_number: { type: string }
                brand: { type: string }
                model: { type: string }
                serial_number: { type: string }
                description: { type: string }
                category: { type: string }
                location: { type: string }
                condition_notes: { type: string }
                purchase_date: { type: string, format: date }
                purchase_price: { type: number }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Tool" } } } }
    delete:
      tags: [Tools]
      summary: Soft-delete a tool (admin only)
      parameters: [{ $ref: "#/components/parameters/ToolId" }]
      responses:
        "200": { description: Deleted }
  /api/tools/{tool_id}/status:
    post:
      tags: [Tools]
      summary: Change tool status directly (admin only — e.g. mark broken)
      parameters: [{ $ref: "#/components/parameters/ToolId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [available, checked_out, broken, under_repair, stolen, lost] }
                notes: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Tool" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/tools/{tool_id}/checkout:
    post:
      tags: [Tools]
      summary: Check out a tool to the current user
      parameters: [{ $ref: "#/components/parameters/ToolId" }]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { notes: { type: string } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Tool" } } } }
        "409": { description: "Tool not available" }
  /api/tools/{tool_id}/checkin:
    post:
      tags: [Tools]
      summary: Check in a checked-out tool
      parameters: [{ $ref: "#/components/parameters/ToolId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes: { type: string }
                condition: { type: string, enum: [available, broken, under_repair, stolen, lost], default: available }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Tool" } } } }
        "409": { description: "Tool is not checked out" }
  /api/tools/{tool_id}/history:
    get:
      tags: [Tools]
      summary: Checkout/status history for one tool
      parameters:
        - { $ref: "#/components/parameters/ToolId" }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
      responses:
        "200": { description: OK }

  # ── Projects ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    userJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    installToken:
      type: http
      scheme: bearer
      description: Per-installation token from POST /api/installations/register
  parameters:
...
    ToolId: { name: tool_id, in: path, required: true, schema: { type: integer } }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }
  schemas:
...
    Tool:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        tool_number: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        model: { type: string, nullable: true }
        serial_number: { type: string, nullable: true }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true }
        location: { type: string, nullable: true }
        status: { type: string, enum: [available, checked_out, broken, under_repair, stolen, lost] }
        status_label: { type: string }
        is_available: { type: boolean }
        hours_checked_out: { type: number, nullable: true }
        is_overdue: { type: boolean }
        condition_notes: { type: string, nullable: true }
        checked_out_by: { type: string, nullable: true }
        checked_out_at: { type: string, format: date-time, nullable: true }
        purchase_date: { type: string, format: date, nullable: true }
        purchase_price: { type: number, nullable: true }
        is_active: { type: boolean }
        barcode_code: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...